Create chemicalTemperatureMonitor
University chemists have developed a new process for the manufacturing of a drug that heals wounds extremely quickly. The manufacturing process is very lengthy and requires monitoring the chemicals at all times, sometimes for hours! Entrusting this task to a student is not possible; students tend to fall asleep or not pay close attention after a while. Therefore I create this program for an automatic device to monitor the manufacturing of the drug. The device measures the temperature every 15 seconds and provides these measurement to your program. 

The program firsts reads two integers representing the minimum and maximum safe temperatures. Next,it continuously reads temperatures (integers) that are being provided by the device. Once the chemical reaction is complete the device will send a value of -999, indicating to you that temperature readings are done. For each recorded temperature that is in the correct range (it could also be equal to the min or max values), the program displays the text "Nothing to report". But as soon as a temperature reaches an unsafe level the program displays the text "Alert!" and stops reading temperatures (although the device may continue sending temperature values).